Change parallax to publico
One day, I got confused about why this was called parallax but parallax
was this other thing and then I was told this was really publico, so
then I was like, "but it says it's parallax" and then some things got
changed, but not all the things. Ergo, this commit.