Remove all the messages for a fingerprint when muted.
I think this is a better behavior for things you don't want to see
(e.g. penises). In the future we should probably add some sort of undo
function for the last mute (accidental mutes seem to happen a lot), in
which case we might need to rethink the removal (keep the muted chats
somewhere so we can bring them back, perhaps?).