[fix] replace broken command in setup
originally the instructions stated that one should run the command ```npm run-script grunt build```.  As there is no npm script named grunt this results in a no-op.

It turns out the the build script is simply aliased to grunt, and as such npm run-script build suffices as a replacement for the broken command in the instruction.

This does raise the question if the build script should even exist.  It could be removed and the developer asked to simply run grunt.